+++ superset-frontend/src/DodoExtensions/components/WarningPanel/styles.ts
@@ -1,29 +1,35 @@
 /* eslint-disable theme-colors/no-literal-colors */
 import { styled } from '@superset-ui/core';
+import Button from '../../../components/Button';
 
 const StyledH4 = styled.h4`
   margin-top: 0;
 `;
 const StyledP = styled.p`
+  hyphens: auto;
   margin-bottom: 10px;
   &:last-child {
     margin-bottom: 0;
   }
 `;
-const WarningPanelWrapper = styled.div`
+const WarningPanelWrapper = styled.div<{ backgroundColor: string }>`
   //height: auto;
   display: flex;
+  background-color: ${props => props.backgroundColor};
   align-items: flex-start;
   justify-content: space-between;
   //box-sizing: border-box;
+
+  border: 1px solid #ffeeba;
+  border-radius: 2px;
 `;
 
 const Alert = styled.div`
   line-height: 22px;
   color: #856404;
-  background-color: #fff3cd;
-  border: 1px solid #ffeeba;
-  border-radius: 2px;
+  //background-color: #fff3cd;
+  //border: 1px solid #ffeeba;
+  //border-radius: 2px;
   padding: 25px;
   width: 100%;
 `;
@@ -36,4 +42,15 @@ const StyledCode = styled.code`
   background-color: #f7f7f7;
 `;
 
-export { StyledH4, StyledP, WarningPanelWrapper, Alert, StyledCode };
+const StyledButton = styled(Button)`
+  padding: 0;
+`;
+
+export {
+  StyledH4,
+  StyledP,
+  WarningPanelWrapper,
+  Alert,
+  StyledCode,
+  StyledButton,
+};
