+++ superset-frontend/src/DodoExtensions/global.ts
@@ -14,3 +14,16 @@ export interface IPanelMsgObj {
   messagesExtra?: string[];
   buttons?: { txt: string; link: string }[];
 }
+
+export interface WarningMsgParams {
+  title?: string;
+  subTitle?: string;
+  body?: string;
+  extra?: string;
+  children?: React.ReactNode;
+  colors?: {
+    backgroundColor: string;
+    textColor: string;
+  };
+  onCLose?: () => void;
+}
